RAG Powered PDF Reader/requirements.txt
streamlit>=1.28.0
pypdf>=3.0.0
faiss-cpu>=1.7.0
numpy>=1.21.0
sentence-transformers>=2.2.0
torch>=2.0.0